+++ src/server/models/user.ts
@@ -110,6 +110,33 @@ class User extends IUser {
     })
   };
 
+  /**
+   * Find the list of user
+   * @returns {Promise<IUser[]>}
+   */
+  static find(): Promise < User[] > {
+    // debug("find");
+    return new Promise < IUser[] >((resolve, reject) => {
+
+      // Do the search
+      _model.find({})
+        .lean()
+        .exec()
+        .then(
+          users => {
+            resolve(users.map(f => {
+              f.id = f._id.toString();
+              return f;
+            }));
+          },
+          err => {
+            reject(err);
+          })
+      ;
+    })
+  }
+
+
   static findById(id: string): Promise < User > {
     return new Promise < IUser >((resolve, reject) => {
       _model.findById(id)
@@ -223,76 +250,4 @@ class User extends IUser {
 }
 
 export = User;
-
-// Init database if it's empty
-User
-  .count()
-  .then(count => {
-    if (count === 0) {
-      // get course ids
-      Course
-        .find()
-        .then(courses => {
-          let fIds = _.shuffle(_.map(courses, f => f['id']))
-          let idNum = 0;
-      var users: {}[] = [{
-        username: 'eric',
-        firstname: 'Eric',
-        lastname: 'Machin',
-        email: 'ermachin@airfance.fr',
-        salt: '6MulWcxSt3p0BC4Xa59Xz9O8TUCu3VJE/cTwT2WOjws5XOnZhPfmj5Rku5EY8xzAgUDAgc5Z/6r1y6JLHtknmGINveNFJhadsAkFE+TS4EqI0Nzm8brPpZ3KZLHVLfF2tDBZtp9K5Z/l5WNOQkCFwEKaSzvbhnP+/i4hCZg2kyk=',
-        hashedPassword: 'b80776ef82727b04f57f314a655a31e926d3ea664e7295445409ef113c04c55984082a076ffbff9df377e19d96667026d498ad9bbd4f6211d6da39d7bb0fae331ad08709928513618174bc5a216ca6b74c5fd8e21cbbae0e7e44e24b6ae1d71a24f4088f10ca3533e144df1d66b9300ace3196097cf933b0a5f234b283b0fce3',
-        isAdmin: true,
-        courses: {}
-      }];
-          users[0]['courses'][fIds[(idNum++) % fIds.length]] = {
-            isFavorite: true,
-            interest: 0.8,
-            dateFollowed: new Date('2016-08-14T00:00:00'),
-            dateFollowedEnd: null,
-            percentFollowed: 0.6,
-          };
-          users[0]['courses'][fIds[(idNum++) % fIds.length]] = {
-            isFavorite: false,
-            interest: 0.2,
-            dateFollowed: new Date('2016-08-22T00:00:00'),
-            dateFollowedEnd: null,
-            percentFollowed: 0.9,
-          };
-          users[0]['courses'][fIds[(idNum++) % fIds.length]] = {
-            isFavorite: false,
-            interest: 0.8,
-            dateFollowed: new Date('2016-09-23T00:00:00'),
-            dateFollowedEnd: new Date('2016-10-01T00:00:00'),
-            percentFollowed: 1,
-          };
-          users[0]['courses'][fIds[(idNum++) % fIds.length]] = {
-            isFavorite: false,
-            interest: 0.1,
-            dateFollowed: new Date('2016-08-12T00:00:00'),
-            dateFollowedEnd: null,
-            percentFollowed: 0.0,
-          };
-          users[0]['courses'][fIds[(idNum++) % fIds.length]] = {
-            isFavorite: false,
-            interest: 0.1,
-            dateFollowed: new Date('2016-10-31T00:00:00'),
-            dateFollowedEnd: new Date('2016-11-11T00:00:00'),
-            percentFollowed: 1,
-          };
-
-          _.forEach(users, o => {
-            var user = new User(o);
-            //console.log(user);
-            User
-              .updateOrCreate(user)
-              .then(user => debug("User created : " + user.username))
-              .catch(err => debug("Error creating user : " + err))
-          });
-        });
-    }
-  })
-  .catch(err => {
-    console.log(err);
-  });
